requirements.txt
matplotlib~=3.6.3
sympy~=1.11.1
discord.py~=2.1.0
topggpy~=1.4.0
psutil~=5.9.4